requirements.txt
Django==3.2
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-filter==2.4.0
django-redis-cache==3.0.0
django-redis==4.12.1
django-stubs==1.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf_yasg==1.20.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
redis==3.5.3
requests==2.25.1
